Refactor user preferences validation and update logic
- Introduced a centralized utility function for validating and processing user preference updates, reducing code duplication across authController and userController.
- Updated the updateUserPreferences functions in both controllers to utilize the new validation utility, improving error handling and ensuring consistent validation of currency and date format.
- Enhanced error responses for invalid inputs and cases where no valid preferences are provided, improving user feedback during updates.
