+++ backend/src/controllers/userController.js
@@ -3,6 +3,7 @@ const User = require('../models/User');
 const apiResponse = require('../utils/apiResponse');
 const { body, validationResult } = require('express-validator');
 const asyncHandler = require('express-async-handler');
+const { validateAndBuildPreferenceUpdate } = require('../utils/preferenceValidation');
 
 /**
  * Validation middleware for profile updates
@@ -127,38 +128,23 @@ exports.getUserPreferences = async (req, res) => {
  */
 exports.updateUserPreferences = async (req, res) => {
   try {
-    const { currency, dateFormat } = req.body;
+    // Use the validation utility to validate and build update object
+    const validationResult = validateAndBuildPreferenceUpdate(req.body);
     
-    // Build preferences update object with only provided fields
-    const preferencesUpdate = {};
-    
-    if (currency) {
-      // Validate currency is in allowed list
-      const allowedCurrencies = ['USD', 'IDR', 'EUR', 'GBP', 'JPY', 'CNY', 'AUD', 'CAD', 'SGD', 'MYR'];
-      if (!allowedCurrencies.includes(currency)) {
-        return apiResponse.badRequest(res, 'Invalid currency');
-      }
-      preferencesUpdate['preferences.currency'] = currency;
+    // If not valid, send error response
+    if (!validationResult.isValid) {
+      return apiResponse.badRequest(res, validationResult.errors[0]);
     }
     
-    if (dateFormat) {
-      // Validate dateFormat is in allowed list
-      const allowedDateFormats = ['DD/MM/YYYY', 'MM/DD/YYYY', 'YYYY-MM-DD'];
-      if (!allowedDateFormats.includes(dateFormat)) {
-        return apiResponse.badRequest(res, 'Invalid date format');
-      }
-      preferencesUpdate['preferences.dateFormat'] = dateFormat;
-    }
-    
-    // If no valid preferences were provided
-    if (Object.keys(preferencesUpdate).length === 0) {
+    // If no updates, send error
+    if (!validationResult.hasUpdates) {
       return apiResponse.badRequest(res, 'No valid preferences provided');
     }
     
     // Update user preferences
     const user = await User.findByIdAndUpdate(
       req.user.id,
-      { $set: preferencesUpdate },
+      { $set: validationResult.preferencesUpdate },
       { new: true, runValidators: true }
     ).select('preferences');
